pyproject.toml
[project]
name = "hdd-lifetime-prediction"
description = "Predicting HDD lifetime using machine learning"
license = {text = "MIT License"}
authors = [
    {name = "Andy Wong", email = "awong36@gmail.com"}
]
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "numpy",
    "pyyaml",
    "flask"
]

[project.optional-dependencies]
dev = [
    "beautifulsoup4",
    "jupyter[lab]",
    "jupytext",
]

[tool.setuptools.packages]
find = {where = ["src"]}

[tool.setuptools.package-data]
"*" = ["*.yaml"]
